+++ internal/repository/interfaces.go
@@ -264,22 +264,22 @@ type MasterDataProcessRepository interface {
 	GetByDate(ctx context.Context, processDate time.Time) (*domain.MasterDataProcess, error)
 
 	// GetByID retrieves a process by its ID
-	GetByID(ctx context.Context, processID int) (*domain.MasterDataProcess, error)
+	GetByID(ctx context.Context, processID int64) (*domain.MasterDataProcess, error)
 
 	// UpdateStatus updates the status of a process
-	UpdateStatus(ctx context.Context, processID int, status string) error
+	UpdateStatus(ctx context.Context, processID int64, status string) error
 
 	// CompleteProcess marks a process as completed
-	CompleteProcess(ctx context.Context, processID int) error
+	CompleteProcess(ctx context.Context, processID int64) error
 
 	// CreateStep creates a new step for a process
-	CreateStep(ctx context.Context, processID int, stepNumber int, stepName string) error
+	CreateStep(ctx context.Context, processID int64, stepNumber int, stepName string) error
 
 	// GetStep retrieves a step by process ID and step number
-	GetStep(ctx context.Context, processID int, stepNumber int) (*domain.MasterDataProcessStep, error)
+	GetStep(ctx context.Context, processID int64, stepNumber int) (*domain.MasterDataProcessStep, error)
 
 	// UpdateStepStatus updates the status of a step
-	UpdateStepStatus(ctx context.Context, processID int, stepNumber int, status string, errorMessage ...string) error
+	UpdateStepStatus(ctx context.Context, processID int64, stepNumber int, status string, errorMessage ...string) error
 
 	// GetFilteredStocks retrieves filtered stocks for a specific date
 	GetFilteredStocks(ctx context.Context, processDate time.Time) ([]domain.FilteredStockRecord, error)
